+++ matcher/utils.ts
@@ -14,6 +14,14 @@ function contains(array: unknown[], value: unknown): boolean {
   return array.some((v) => equal(v, value));
 }
 
+function containSome(target: unknown[], part: unknown[]): boolean {
+  return part.some((value) => contains(target, value));
+}
+
+function containAll(target: unknown[], part: unknown[]): boolean {
+  return part.every((value) => contains(target, value));
+}
+
 function has(
   key: PropertyKey,
   object: object,
@@ -62,4 +70,13 @@ function hasPath(
   return false;
 }
 
-export { contains, has, hasPath, propPath, takeLast };
+export {
+  containAll,
+  contains,
+  containSome,
+  has,
+  hasPath,
+  prop,
+  propPath,
+  takeLast,
+};